requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
